refactor(runtime): extract post-execution recording into agent/turn — #7415
The post-execution loop moves to
turn/post_exec.rs::record_executed_outcomes: per-outcome
tool_call_result log, after_tool_call hook, completion Status to the
draft, and the executed calls' ordered_results slots. ordered_results
stays owned by the loop skeleton, passed &mut.
