+++ crates/zeroclaw-runtime/src/agent/loop_.rs
@@ -1116,68 +1116,15 @@ pub async fn run_tool_call_loop(
             .await?
         };
 
-        for ((idx, call), outcome) in executable_indices
-            .iter()
-            .zip(executable_calls.iter())
-            .zip(executed_outcomes)
-        {
-            ::zeroclaw_log::record!(
-                INFO,
-                ::zeroclaw_log::Event::new(module_path!(), ::zeroclaw_log::Action::Complete)
-                    .with_outcome(if outcome.success {
-                        ::zeroclaw_log::EventOutcome::Success
-                    } else {
-                        ::zeroclaw_log::EventOutcome::Failure
-                    })
-                    .with_duration(u64::try_from(outcome.duration.as_millis()).unwrap_or(u64::MAX),)
-                    .with_attrs(::serde_json::json!({
-                        "model": model,
-                        "iteration": iteration + 1,
-                        "tool": call.name.clone(),
-                        "error_reason": outcome.error_reason,
-                        "output": scrub_credentials(&outcome.output),
-                        "trace_id": turn_id,
-                    })),
-                "tool_call_result"
-            );
-
-            // ── Hook: after_tool_call (void) ─────────────────
-            if let Some(hooks) = hooks {
-                let tool_result_obj = crate::tools::ToolResult {
-                    success: outcome.success,
-                    output: outcome.output.clone(),
-                    error: None,
-                };
-                hooks
-                    .fire_after_tool_call(&call.name, &tool_result_obj, outcome.duration)
-                    .await;
-            }
-
-            // ── Progress: tool completion ───────────────────────
-            if let Some(ref tx) = on_delta {
-                let secs = outcome.duration.as_secs();
-                let progress_msg = if outcome.success {
-                    format!("\u{2705} {} ({secs}s)\n", call.name)
-                } else if let Some(ref reason) = outcome.error_reason {
-                    format!(
-                        "\u{274c} {} ({secs}s): {}\n",
-                        call.name,
-                        truncate_with_ellipsis(reason, 200)
-                    )
-                } else {
-                    format!("\u{274c} {} ({secs}s)\n", call.name)
-                };
-                ::zeroclaw_log::record!(
-                    DEBUG,
-                    ::zeroclaw_log::Event::new(module_path!(), ::zeroclaw_log::Action::Note)
-                        .with_attrs(::serde_json::json!({"tool": call.name, "secs": secs})),
-                    "Sending progress complete to draft"
-                );
-                let _ = tx.send(StreamDelta::Status(progress_msg)).await;
-            }
-
-            ordered_results[*idx] = Some((call.name.clone(), call.tool_call_id.clone(), outcome));
-        }
+        super::turn::record_executed_outcomes(
+            &ctx,
+            &executable_indices,
+            &executable_calls,
+            executed_outcomes,
+            &mut ordered_results,
+            iteration,
+        )
+        .await;
 
         // Collect tool results and build per-tool output for loop detection.
         // Only non-ignored tool outputs contribute to the identical-output hash.
